Added relationships, migrations and models

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -21,7 +21,7 @@ class CreateUsersTable extends Migration
             $table->string('email')->unique();
             $table->timestamp('email_verified_at')->nullable();
             $table->string('password');
-            $table->integer('condition_id')->nullable();
+            $table->unsignedBigInteger('condition_id')->nullable();
             $table->rememberToken();
             $table->timestamps();
         });
