+++ database/migrations/2024_11_21_115812_create_projects_table.php
@@ -17,8 +17,8 @@ class CreateProjectsTable extends Migration
             $table->id();
             $table->string('name');
             $table->text('description')->nullable();
-            $table->integer('condition_id')->nullable();
-            $table->integer('status_id')->nullable();
+            $table->unsignedBigInteger('condition_id')->nullable();
+            $table->unsignedBigInteger('status_id')->nullable();
             $table->timestamps();
         });
     }
