+++ database/migrations/2024_11_21_120027_create_project_task_table.php
@@ -15,8 +15,8 @@ class CreateProjectTaskTable extends Migration
     {
         Schema::create('project_task', function (Blueprint $table) {
             $table->id();
-            $table->integer('project_id');
-            $table->integer('task_id');
+            $table->unsignedBigInteger('project_id');
+            $table->unsignedBigInteger('task_id');
             $table->timestamps();
         });
     }
